test: move data usecase tests to transaction's folder
