+++ tests/data/usecases/transaction/db-find-transaction-by-id.spec.ts
@@ -1,6 +1,6 @@
-import { DbFindTransactionById } from "../../../src/data/usecases/transaction/db-find-transaction-by-id";
-import { FindTransactionByIdRepository } from "../../../src/data/protocols/db/transaction/find-transaction-by-id-repository";
-import { TransactionModel } from "../../../src/domain/models/transaction";
+import { DbFindTransactionById } from "../../../../src/data/usecases/transaction/db-find-transaction-by-id";
+import { FindTransactionByIdRepository } from "../../../../src/data/protocols/db/transaction/find-transaction-by-id-repository";
+import { TransactionModel } from "../../../../src/domain/models/transaction";
 
 const makeSut = (): {
   sut: DbFindTransactionById;
