+++ tests/data/usecases/transaction/db-revert-transaction.spec.ts
@@ -1,5 +1,5 @@
-import { DbRevertTransaction } from "../../../src/data/usecases/transaction/db-revert-transaction";
-import { RevertTransactionRepository } from "../../../src/data/protocols/db/transaction/revert-transaction-repository";
+import { DbRevertTransaction } from "../../../../src/data/usecases/transaction/db-revert-transaction";
+import { RevertTransactionRepository } from "../../../../src/data/protocols/db/transaction/revert-transaction-repository";
 
 const makeSut = (): {
   sut: DbRevertTransaction;
